moved map and components to unified folder
