+++ webapp/src/pages/Dashboard.js
@@ -5,7 +5,7 @@ import Button from "react-bootstrap/Button";
 import Col from "react-bootstrap/Col";
 import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
-import Map from "./Map";
+import Map from "../components/mobilitymap/Map";
 
 const Dashboard = () => {
   const { user } = useUser();
